refactor: move switzer font import to ui/fonts and clean up imports
Consolidate the switzer font import from layout to a centralized location in ui/fonts to improve maintainability. Additionally, clean up unused imports in the settings page to reduce clutter and improve code readability.
